# Conflicts:
#	driver-core/src/test/functional/com/mongodb/internal/connection/TestCommandListener.java
#	driver-sync/src/test/functional/com/mongodb/client/AbstractClientSideOperationsTimeoutProseTest.java
#	driver-sync/src/test/functional/com/mongodb/client/ClientSideOperationTimeoutProseTest.java

+++ driver-sync/src/test/functional/com/mongodb/client/unified/EventMatcher.java
@@ -99,7 +99,7 @@ final class EventMatcher {
 
             if (expected.containsKey("hasServerConnectionId")) {
                 boolean hasServerConnectionId = expected.getBoolean("hasServerConnectionId").getValue();
-                Integer serverConnectionId = actual.getConnectionDescription().getConnectionId().getServerValue();
+                Long serverConnectionId = actual.getConnectionDescription().getConnectionId().getServerValue();
                 if (hasServerConnectionId) {
                     assertNotNull(context.getMessage("Expected serverConnectionId"), serverConnectionId);
                 } else {
